Web Scraper/requirements.txt
pillow
requests
beautifulsoup4
pandas
